script
Former-commit-id: adfd322ea5fb178ec83df7610cf46a3ec6148ee2
Former-commit-id: 484b5bab1e9fdc4bcfe2bf2c24eeec3cf7e5fdb9
Former-commit-id: a64e4d8e5f2c559afdd3132322417c035d14209d [formerly 10cff0085863c5663e9f4a1c2e9f747aa1926273] [formerly 510213f266c4b0e22e95591af5f9cd026996c238 [formerly 895e61ba16bcc79e8e5b04c18ff8ba54b39f7d35]]
Former-commit-id: 5c5260538960501b8fce9fb2d47efcb8e31a8fdb [formerly 128a4d613002bd824daad8de8fa5ec0fba87d985]
Former-commit-id: d1521153e53efc45070143ce6bc19d5cb8b297c7
Former-commit-id: 909095e134bf3e3c298a595a27bdfe28d044bb95
Former-commit-id: ffb5d68cf3bcb866465f65c108ac04d2d5a3170b

+++ vnncomp_scripts/install_tool.sh
@@ -10,9 +10,6 @@ NEURALSAT_PY=$NEURALSAT_CONDA_HOME/bin/python
 NEURALSAT_HOME=$(dirname $(pwd))
 NEURALSAT_MAIN=$NEURALSAT_HOME/neuralsat/main.py
 
-echo "Neuralsat home is                           $NEURALSAT_HOME"
-echo "Conda will be installed to                  $CONDA_HOME"
-echo "Neuralsat's conda env will be installed at  $NEURALSAT_CONDA_HOME"
 
 # check arguments
 if [ "$1" != ${VERSION_STRING} ]; then
@@ -22,6 +19,10 @@ fi
 
 
 echo "======= Installing $TOOL_NAME ======="
+echo "Neuralsat home is                           $NEURALSAT_HOME"
+echo "Conda will be installed to                  $CONDA_HOME"
+echo "Neuralsat's conda env will be installed at  $NEURALSAT_CONDA_HOME"
+
 
 
 wget -O conda.sh https://repo.anaconda.com/miniconda/Miniconda3-py310_23.3.1-0-Linux-x86_64.sh
@@ -40,21 +41,37 @@ $CONDA  env create -p $NEURALSAT_CONDA_HOME -f ../env.yaml
 
 
 # setup alias
-echo "======= Adding alias ======="
-echo "neuralsat alias added"
+echo "======= Exporting env variables ======="
 alias neuralsat='$NEURALSAT_PY $NEURALSAT_MAIN'
-sed -i '/neuralsat/d' ~/.bashrc
+sed -i '/neuralsat/Id' ~/.bashrc
+echo "####         Neuralsat stuff            ####" >> ~/.bashrc
 echo "alias neuralsat='$NEURALSAT_PY $NEURALSAT_MAIN'" >> ~/.bashrc
 
+export NEURALSAT_CONDA_HOME
+export NEURALSAT_HOME
+export NEURALSAT_PY
+export NEURALSAT_MAIN
+echo "export NEURALSAT_CONDA_HOME=$NEURALSAT_CONDA_HOME" >> ~/.bashrc
+echo "export NEURALSAT_HOME=$NEURALSAT_HOME" >>  ~/.bashrc
+echo "export NEURALSAT_PY=$NEURALSAT_PY" >> ~/.bashrc
+echo "export NEURALSAT_MAIN=$NEURALSAT_MAIN" >> ~/.bashrc
+
+echo "####       End of Neuralsat stuff       ####" >> ~/.bashrc
 
 # install gurobi license
 echo "======= Installing Gurobi License ======="
-echo "enter WLSACCESSID"
-read ACCESSID
-echo "enter WLSSECRET"
-read SECRET
-echo "enter LICENSEID"
-read LICENSEID
-echo "WLSACCESSID=$ACCESSID" > ~/gurobi.lic
-echo "WLSSECRET=$SECRET" >> ~/gurobi.lic
-echo "LICENSEID=$LICENSEID" >> ~/gurobi.lic
+if [ -e ~/gurobi.lic ]; then
+    echo "Gurobi license detected ~/gurobi.lic"
+else
+    echo "No Gurobi license found...enter it now"
+    echo "enter WLSACCESSID"
+    read ACCESSID
+    echo "enter WLSSECRET"
+    read SECRET
+    echo "enter LICENSEID"
+    read LICENSEID
+    echo "WLSACCESSID=$ACCESSID" > ~/gurobi.lic
+    echo "WLSSECRET=$SECRET" >> ~/gurobi.lic
+    echo "LICENSEID=$LICENSEID" >> ~/gurobi.lic
+fi
+
